requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.5
dataclasses-json==0.5.7
docx==0.2.4
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.177
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
pandas==2.0.1
Pillow==9.5.0
pydantic==1.10.7
pydub==0.25.1
python-docx==0.8.11
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
SQLAlchemy==2.0.12
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
XlsxWriter==3.1.0
yarl==1.8.2
zipp==3.15.0
